add query parameter
